fix(gdn-prefill): capability-gate the ldmab kernel + split FP8 ops out of gemm_dense
fp8_fp8_gemm_ldmab is only compiled for the qwen3.6-27b/nvfp4 target, but
fp8_gemm_n128 is shared by every model on the FP8 prefill path (35b-a3b,
qwen3-next-80b, gemma-4, minimax, deepseek-v4, ...). The default-on routing
looked the kernel up with .expect(), which would panic at first prefill on all
of those targets. Probe the handles with .ok() instead and fall through to the
validated scalar fp8_gemm_t when the kernel is absent.

Also splits the FP8 dense-GEMM ops into ops/gemm_dense_fp8.rs (sibling of
gemm_dense_int8.rs) — gemm_dense.rs had grown to 521 LoC, over the 500 cap —
and fixes clippy (manual_is_multiple_of) + rustfmt.
