+++ crates/spark-model/examples/fp8gemm_microtest.rs
@@ -132,58 +132,116 @@ fn main() -> Result<()> {
     // A/B probe: time fp8_fp8_gemm_t (FP8 activation, no in-loop convert) vs
     // fp8_gemm_t (BF16 activation). Accuracy-neutral routing candidate.
     if std::env::var_os("ATLAS_PROBE_FP8FP8").is_some() {
-        let a_fp8: Vec<u8> = (0..m * k).map(|i| f32_to_e4m3(bf16_bits_to_f32(a_bf16[i]))).collect();
+        let a_fp8: Vec<u8> = (0..m * k)
+            .map(|i| f32_to_e4m3(bf16_bits_to_f32(a_bf16[i])))
+            .collect();
         let a8_ptr = upload_bytes(gpu, &a_fp8)?;
         // ldmab: FP8 A, FP8 B, grid (N/128, M/128), block 256 — vs scalar-load fp8_gemm_t.
         {
             let h = gpu.kernel("w4a16", "fp8_fp8_gemm_ldmab")?;
-            let launch = |s| KernelLaunch::new(gpu, h)
-                .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 128), 1]).block([256, 1, 1])
-                .arg_ptr(a8_ptr).arg_ptr(b_ptr).arg_ptr(c_ptr)
-                .arg_u32(m as u32).arg_u32(n as u32).arg_u32(k as u32).launch(s);
-            for _ in 0..8 { launch(stream)?; }
+            let launch = |s| {
+                KernelLaunch::new(gpu, h)
+                    .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 128), 1])
+                    .block([256, 1, 1])
+                    .arg_ptr(a8_ptr)
+                    .arg_ptr(b_ptr)
+                    .arg_ptr(c_ptr)
+                    .arg_u32(m as u32)
+                    .arg_u32(n as u32)
+                    .arg_u32(k as u32)
+                    .launch(s)
+            };
+            for _ in 0..8 {
+                launch(stream)?;
+            }
             gpu.synchronize(stream)?;
             let iters = 60u32;
             let t0 = std::time::Instant::now();
-            for _ in 0..iters { launch(stream)?; }
+            for _ in 0..iters {
+                launch(stream)?;
+            }
             gpu.synchronize(stream)?;
             let secs = t0.elapsed().as_secs_f64() / iters as f64;
             let flop = 2.0 * m as f64 * n as f64 * k as f64;
-            println!("PROBE fp8_fp8_gemm_ldmab: M={m} N={n} K={k} {:.3} ms {:.1} TFLOP/s", secs * 1e3, flop / secs / 1e12);
+            println!(
+                "PROBE fp8_fp8_gemm_ldmab: M={m} N={n} K={k} {:.3} ms {:.1} TFLOP/s",
+                secs * 1e3,
+                flop / secs / 1e12
+            );
             // CORRECTNESS: ldmab vs fp8_fp8_gemm_t (both fp8xfp8, same math) — cosine ~1.0 iff layout correct.
-            launch(stream)?; gpu.synchronize(stream)?;
+            launch(stream)?;
+            gpu.synchronize(stream)?;
             let mut craw = vec![0u8; m * n * 2];
             gpu.copy_d2h(c_ptr, &mut craw)?;
-            let cldm: Vec<f32> = craw.chunks_exact(2).map(|c| f32::from_bits((u16::from_le_bytes([c[0], c[1]]) as u32) << 16)).collect();
+            let cldm: Vec<f32> = craw
+                .chunks_exact(2)
+                .map(|c| f32::from_bits((u16::from_le_bytes([c[0], c[1]]) as u32) << 16))
+                .collect();
             let hff = gpu.kernel("w4a16", "fp8_fp8_gemm_t")?;
-            KernelLaunch::new(gpu, hff).grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1]).block([128, 1, 1])
-                .arg_ptr(a8_ptr).arg_ptr(b_ptr).arg_ptr(c_ptr)
-                .arg_u32(m as u32).arg_u32(n as u32).arg_u32(k as u32).launch(stream)?;
+            KernelLaunch::new(gpu, hff)
+                .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1])
+                .block([128, 1, 1])
+                .arg_ptr(a8_ptr)
+                .arg_ptr(b_ptr)
+                .arg_ptr(c_ptr)
+                .arg_u32(m as u32)
+                .arg_u32(n as u32)
+                .arg_u32(k as u32)
+                .launch(stream)?;
             gpu.synchronize(stream)?;
             gpu.copy_d2h(c_ptr, &mut craw)?;
-            let cff: Vec<f32> = craw.chunks_exact(2).map(|c| f32::from_bits((u16::from_le_bytes([c[0], c[1]]) as u32) << 16)).collect();
+            let cff: Vec<f32> = craw
+                .chunks_exact(2)
+                .map(|c| f32::from_bits((u16::from_le_bytes([c[0], c[1]]) as u32) << 16))
+                .collect();
             let (mut d, mut na, mut nb) = (0f64, 0f64, 0f64);
-            for i in 0..cldm.len() { let (x, y) = (cldm[i] as f64, cff[i] as f64); d += x*y; na += x*x; nb += y*y; }
+            for i in 0..cldm.len() {
+                let (x, y) = (cldm[i] as f64, cff[i] as f64);
+                d += x * y;
+                na += x * x;
+                nb += y * y;
+            }
             let cos = d / (na.sqrt() * nb.sqrt() + 1e-30);
-            println!("PROBE ldmab_vs_fp8fp8 cosine={cos:.6}  ({}=PASS)", if cos >= 0.99 { "OK" } else { "FAIL" });
+            println!(
+                "PROBE ldmab_vs_fp8fp8 cosine={cos:.6}  ({}=PASS)",
+                if cos >= 0.99 { "OK" } else { "FAIL" }
+            );
         }
         for (name, aptr) in [("fp8_fp8_gemm_t", a8_ptr), ("fp8_gemm_t", a_ptr)] {
             let h = gpu.kernel("w4a16", name)?;
-            let launch = |s| KernelLaunch::new(gpu, h)
-                .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1]).block([128, 1, 1])
-                .arg_ptr(aptr).arg_ptr(b_ptr).arg_ptr(c_ptr)
-                .arg_u32(m as u32).arg_u32(n as u32).arg_u32(k as u32).launch(s);
-            for _ in 0..8 { launch(stream)?; }
+            let launch = |s| {
+                KernelLaunch::new(gpu, h)
+                    .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1])
+                    .block([128, 1, 1])
+                    .arg_ptr(aptr)
+                    .arg_ptr(b_ptr)
+                    .arg_ptr(c_ptr)
+                    .arg_u32(m as u32)
+                    .arg_u32(n as u32)
+                    .arg_u32(k as u32)
+                    .launch(s)
+            };
+            for _ in 0..8 {
+                launch(stream)?;
+            }
             gpu.synchronize(stream)?;
             let iters = 60u32;
             let t0 = std::time::Instant::now();
-            for _ in 0..iters { launch(stream)?; }
+            for _ in 0..iters {
+                launch(stream)?;
+            }
             gpu.synchronize(stream)?;
             let secs = t0.elapsed().as_secs_f64() / iters as f64;
             let flop = 2.0 * m as f64 * n as f64 * k as f64;
-            println!("PROBE {name}: M={m} N={n} K={k} {:.3} ms {:.1} TFLOP/s", secs * 1e3, flop / secs / 1e12);
+            println!(
+                "PROBE {name}: M={m} N={n} K={k} {:.3} ms {:.1} TFLOP/s",
+                secs * 1e3,
+                flop / secs / 1e12
+            );
         }
-        for p in [a8_ptr] { gpu.free(p).ok(); }
+        for p in [a8_ptr] {
+            gpu.free(p).ok();
+        }
     }
 
     let handle = gpu.kernel("w4a16", "fp8_gemm_t")?;
@@ -205,8 +263,12 @@ fn main() -> Result<()> {
             KernelLaunch::new(gpu, handle)
                 .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1])
                 .block([128, 1, 1])
-                .arg_ptr(a_ptr).arg_ptr(b_ptr).arg_ptr(c_ptr)
-                .arg_u32(m as u32).arg_u32(n as u32).arg_u32(k as u32)
+                .arg_ptr(a_ptr)
+                .arg_ptr(b_ptr)
+                .arg_ptr(c_ptr)
+                .arg_u32(m as u32)
+                .arg_u32(n as u32)
+                .arg_u32(k as u32)
                 .launch(stream)?;
         }
         gpu.synchronize(stream)?;
@@ -216,14 +278,22 @@ fn main() -> Result<()> {
             KernelLaunch::new(gpu, handle)
                 .grid([div_ceil(n as u32, 128), div_ceil(m as u32, 64), 1])
                 .block([128, 1, 1])
-                .arg_ptr(a_ptr).arg_ptr(b_ptr).arg_ptr(c_ptr)
-                .arg_u32(m as u32).arg_u32(n as u32).arg_u32(k as u32)
+                .arg_ptr(a_ptr)
+                .arg_ptr(b_ptr)
+                .arg_ptr(c_ptr)
+                .arg_u32(m as u32)
+                .arg_u32(n as u32)
+                .arg_u32(k as u32)
                 .launch(stream)?;
         }
         gpu.synchronize(stream)?;
         let secs = t0.elapsed().as_secs_f64() / iters as f64;
         let flop = 2.0 * m as f64 * n as f64 * k as f64;
-        println!("TIMING: M={m} N={n} K={k} {:.3} ms/iter {:.1} TFLOP/s", secs * 1e3, flop / secs / 1e12);
+        println!(
+            "TIMING: M={m} N={n} K={k} {:.3} ms/iter {:.1} TFLOP/s",
+            secs * 1e3,
+            flop / secs / 1e12
+        );
     }
 
     let mut c_raw = vec![0u8; m * n * 2];
